moving the logic of reset password to the user controller
