+++ routes/api.php
@@ -5,7 +5,6 @@ use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\api\UserController;
 use App\Http\Controllers\api\FaqController;
 use App\Http\Controllers\api\AuthController;
-use App\Http\Controllers\api\ForgetController;
 use App\Http\Controllers\api\CodeController;
 use App\Http\Controllers\api\AppController;
 
@@ -24,8 +23,8 @@ Route::post('/register', [UserController::class, 'store'])->name('register');
 Route::post('/login', [AuthController::class, 'login'])->name('login');
 
 // recovery password
-Route::post('/forgetpassword', [ForgetController::class, 'forgetPassword'])->name('forgetpassword');
-Route::post('/resetPassword/{token}', [ForgetController::class, 'resetPassword'])->name('resetPassword');
+Route::post('/forgetpassword', [UserController::class, 'forgetPassword'])->name('forget.password');
+Route::post('/resetPassword/{token}', [UserController::class, 'resetPassword'])->name('reset.password');
 
 Route::get('/faqs', [FaqController::class, 'index']);
 Route::get('/apps', [AppController::class, 'index'])->name('app.index');
